config: separate schema code and data validators
This patch separates the instance config schema and data validators.
The validation logic is moved to `box/lua/config/validators.lua`,
improving schema readability by separating validation annotations
from the schema structure.

Closes tarantool#10691

NO_TEST=Refactoring
NO_DOC=Refactoring
NO_CHANGELOG=Refactoring
