+++ src/box/CMakeLists.txt
@@ -64,6 +64,7 @@ lua_source(lua_sources lua/config/init.lua                 config_init_lua)
 lua_source(lua_sources lua/config/instance_config.lua      config_instance_config_lua)
 lua_source(lua_sources lua/config/source/env.lua           config_source_env_lua)
 lua_source(lua_sources lua/config/source/file.lua          config_source_file_lua)
+lua_source(lua_sources lua/config/validators.lua           config_validators_lua)
 lua_source(lua_sources lua/config/utils/aboard.lua         config_utils_aboard_lua)
 lua_source(lua_sources lua/config/utils/expression.lua     config_utils_expression_lua)
 lua_source(lua_sources lua/config/utils/file.lua           config_utils_file_lua)
